more warning fixes
git-svn-id: http://svn.miranda-ng.org/main/trunk@11183 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ src/core/stdclist/src/clcpaint.cpp
@@ -484,7 +484,7 @@ void PaintClc(HWND hwnd, struct ClcData *dat, HDC hdc, RECT * rcPaint)
 					rc.right = (clRect.right - clRect.left);
 					rc.bottom = rc.top;
 					if (qlen)
-						DrawText(hdcMem, szText, qlen, &rc, DT_EDITCONTROL | DT_NOPREFIX | DT_NOCLIP | DT_WORD_ELLIPSIS | DT_SINGLELINE);
+						DrawText(hdcMem, szText, (int)qlen, &rc, DT_EDITCONTROL | DT_NOPREFIX | DT_NOCLIP | DT_WORD_ELLIPSIS | DT_SINGLELINE);
 				}
 			}
 		}
